Lots of Imgui changes
- Collider Components
- Font Manager
- Text Update
- Map Generator
- Content Generator

+++ src/Twin2Engine/GraphicEngine/src/graphic/manager/TextureManager.cpp
@@ -95,7 +95,7 @@ Texture2D* TextureManager::LoadTexture2D(const string& path, const TextureFileFo
         return nullptr;
     }
 
-    glTexImage2D(GL_TEXTURE_2D, 0, internalFormat, width, height, 0, format, GL_UNSIGNED_BYTE, imgData);
+    glTexImage2D(GL_TEXTURE_2D, 0, (GLint)internalFormat, width, height, 0, format, GL_UNSIGNED_BYTE, imgData);
     glGenerateMipmap(GL_TEXTURE_2D);
     stbi_image_free(imgData);
